feat(web): /settings 3-section IA
Splits /settings into three child sections: Account, Preferences,
Materials. Adds a left-rail (desktop) / horizontal-scroll (mobile)
section nav inside SettingsView, with Materials as the default so
existing entry points still land on the per-material tabs. Account
and Preferences are empty placeholders for now — Account fills with
the migrated #92 chores next commit; Preferences stays empty at
launch per the spec.

The existing year-tabs + year-card UI moves verbatim into
SettingsMaterialsView so the refactor is purely structural; per-club
chain UI replaces the scope ladder in a later commit.
